libcava + cava application
