+++ .github/workflows/build-and-test.yml
@@ -39,7 +39,7 @@ jobs:
     - name: build cavacore test application
       run: gcc -c -g cavacore_test.c
     - name: link cavacore test application
-      run: gcc -o cavacore_test cavacore_test.o cava-cavacore.o -lm -lfftw3
+      run: gcc -o cavacore_test cavacore_test.o src/cava-cavacore.o -lm -lfftw3
     - name: run cavacore test application
       run: ./cavacore_test
