+++ include/common.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "cavacore.h"
+#include "config.h"
+#include "input/common.h"
+#include "output/common.h"
+
+void cleanup(int output_mode);
