+++ meson_options.txt
@@ -1,0 +1,11 @@
+option('input_alsa', type: 'feature', value: 'auto', description: 'build with support for ALSA')
+option('input_portaudio', type: 'feature', value: 'auto', description: 'build with support for portaudio')
+option('input_pulse', type: 'feature', value: 'auto', description: 'build with support for Pulse audio')
+option('asan', type: 'feature', value: 'auto', description: 'build with AddressSanitizer')
+option('tsan', type: 'feature', value: 'auto', description: 'build with ThreadSanitizer')
+option('ubsan', type: 'feature', value: 'auto', description: 'build with UndefinedBehaviorSanitizer')
+option('input_sndio', type: 'feature', value: 'auto', description: 'build with support for input from sndio')
+option('input_pipewire', type: 'feature', value: 'auto', description: 'build with support for input from pipewire')
+option('output_sdl', type: 'feature', value: 'auto', description: 'build with support for output to sdl')
+option('output_ncurses', type: 'feature', value: 'auto', description: 'build with support for output to ncurses')
+option('output_sdl_glsl', type: 'feature', value: 'auto', description: 'build with support for output to sdl_glsl')
