+++ src/output/noritake.c
@@ -1,4 +1,3 @@
-#include "../config.h"
 #include <math.h>
 #include <stdint.h>
 #include <stdio.h>
