+++ src/output/shaders/normalized_bars.frag
