+++ src/output/terminal_bcircle.c
